Moved block response logic to separate mixin
