+++ tests/integration/test_backbone.py
@@ -1,10 +1,12 @@
-from typing import Dict, Any, Optional
+from typing import Dict
 
 import pytest
 
 from bami.backbone.block import BamiBlock
-from bami.backbone.community import BlockResponse, BamiCommunity
+from bami.backbone.blockresponse import BlockResponseMixin, BlockResponse
+from bami.backbone.community import BamiCommunity
 from bami.backbone.sub_community import BaseSubCommunity, LightSubCommunity
+from bami.backbone.utils import decode_raw, encode_raw
 from ipv8.keyvault.crypto import default_eccrypto
 from tests.mocking.base import (
     SetupValues,
@@ -19,17 +21,6 @@ class SimpleCommunity(BamiCommunity):
     A very basic community with no additional functionality. Used during the integration tests.
     """
 
-    def apply_confirm_tx(self, block: BamiBlock, confirm_tx: Dict) -> None:
-        pass
-
-    def apply_reject_tx(self, block: BamiBlock, reject_tx: Dict) -> None:
-        pass
-
-    def block_response(
-        self, block: BamiBlock, wait_time: float = None, wait_blocks: int = None
-    ) -> BlockResponse:
-        return BlockResponse.DELAY
-
     def create_subcom(self, *args, **kwargs) -> BaseSubCommunity:
         return LightSubCommunity(*args, **kwargs)
 
@@ -37,18 +28,40 @@ class SimpleCommunity(BamiCommunity):
         pass
 
 
-NUM_NODES = 2
+class BlockResponseCommunity(BlockResponseMixin, SimpleCommunity):
+    """
+    Basic community with block response functionality enabled.
+    """
+
+    def apply_confirm_tx(self, block: BamiBlock, confirm_tx: Dict) -> None:
+        pass
+
+    def apply_reject_tx(self, block: BamiBlock, reject_tx: Dict) -> None:
+        pass
+
+    def received_block_in_order(self, block: BamiBlock) -> None:
+        print(block.transaction)
+        decoded_tx = decode_raw(block.transaction)
+        if decoded_tx.get(b"to_peer", None) == self.my_peer.public_key.key_to_bin():
+            self.add_block_to_response_processing(block)
+
+    def block_response(
+        self, block: BamiBlock, wait_time: float = None, wait_blocks: int = None
+    ) -> BlockResponse:
+        if block.type == b"good":
+            return BlockResponse.CONFIRM
+        elif block.type == b"bad":
+            return BlockResponse.REJECT
+        return BlockResponse.DELAY
 
 
 @pytest.fixture
-async def set_vals(tmpdir_factory):
+async def set_vals(tmpdir_factory, community_cls, num_nodes):
     dirs = [
         tmpdir_factory.mktemp(str(SimpleCommunity.__name__), numbered=True)
-        for _ in range(NUM_NODES)
+        for _ in range(num_nodes)
     ]
-    nodes = create_and_connect_nodes(
-        NUM_NODES, work_dirs=dirs, ov_class=SimpleCommunity
-    )
+    nodes = create_and_connect_nodes(num_nodes, work_dirs=dirs, ov_class=community_cls)
     # Make sure every node has a community to listen to
     community_key = default_eccrypto.generate_key("curve25519").pub()
     community_id = community_key.key_to_bin()
@@ -61,6 +74,8 @@ async def set_vals(tmpdir_factory):
 
 
 @pytest.mark.asyncio
+@pytest.mark.parametrize("community_cls", [SimpleCommunity])
+@pytest.mark.parametrize("num_nodes", [2])
 async def test_simple_frontier_reconciliation_after_partition(set_vals):
     """
     Test whether missing blocks are synchronized after a network partition.
@@ -84,3 +99,19 @@ async def test_simple_frontier_reconciliation_after_partition(set_vals):
     assert len(frontier2.terminal) == 1
     assert frontier2.terminal[0][0] == 3
     assert frontier1 == frontier2
+
+
+@pytest.mark.asyncio
+@pytest.mark.parametrize("community_cls", [BlockResponseCommunity])
+@pytest.mark.parametrize("num_nodes", [2])
+async def test_block_confirm(set_vals):
+    """
+    Test whether blocks are confirmed correctly.
+    """
+    block = set_vals.nodes[0].overlay.create_signed_block(
+        com_id=set_vals.community_id,
+        transaction=encode_raw({b"to_peer": 3}),
+        block_type=b"good",
+    )
+    set_vals.nodes[0].overlay.share_in_community(block, set_vals.community_id)
+    await deliver_messages()
